legal-agent-poc/requirements.txt
pydantic-ai[openai]>=0.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
qdrant-client>=1.12.0
openai>=1.50.0
httpx>=0.27.0
logfire>=2.0.0
